added security function, login and logout page. refactored pages to handle role level CRUD matrix

+++ orders.php
@@ -5,66 +5,77 @@ include_once 'orders_crud.php';
 <!DOCTYPE html>
 <html>
 <head>
- <meta charset="utf-8">
- <meta http-equiv="X-UA-Compatible" content="IE=edge">
- <meta name="viewport" content="width=device-width, initial-scale=1">
- <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
- <title>Harvest Hill: Orders</title>
- <!-- Bootstrap -->
- <link href="css/bootstrap.min.css" rel="stylesheet">
- 
- <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
- <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
-  </head>
-  <body>
-    <?php include_once 'nav_bar.php'; ?>
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <!-- The above 3 meta tags *must* come first in the head; any other head content must come *after* these tags -->
+  <title>Harvest Hill: Orders</title>
+  <!-- Bootstrap -->
+  <link href="css/bootstrap.min.css" rel="stylesheet">
 
-    <div class="container-fluid">
-     <div class="row">
-      <div class="col-xs-12 col-sm-8 col-sm-offset-2 col-md-6 col-md-offset-3">
-        <div class="page-header">
-          <h2>Create New Order</h2>
+  <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
+  <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
+  <!--[if lt IE 9]>
+    <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
+    <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
+  <![endif]-->
+  <?php
+  // We need to use sessions, so you should always start sessions using the below code.
+  session_start();
+  // If the user is not logged in redirect to the login page...
+  if (!isset($_SESSION["loggedin"]) || $_SESSION["loggedin"] !== true && !isset($_SESSION["role"])) {
+    header('Location: login.php');
+    exit;
+  } else {
+    $role = $_SESSION['role'];
+  }
+  ?>
+</head>
+<body>
+  <?php include_once 'nav_bar.php'; ?>
+
+  <div class="container-fluid">
+   <div class="row">
+    <div class="col-xs-12 col-sm-8 col-sm-offset-2 col-md-6 col-md-offset-3">
+      <div class="page-header">
+        <h2>Create New Order</h2>
+      </div>
+      <form action="orders.php" method="post" class="form-horizontal">
+       <div class="form-group">
+        <label for="orderid" class="col-sm-3 control-label">Order ID</label>
+        <div class="col-sm-9">
+          <input name="oid" type="text" class="form-control" id="orderid" placeholder="Order ID" value="<?php if(isset($_GET['edit'])) echo $editrow['fld_order_id']; ?>" required>
         </div>
-        <form action="orders.php" method="post" class="form-horizontal">
-         <div class="form-group">
-          <label for="orderid" class="col-sm-3 control-label">Order ID</label>
-          <div class="col-sm-9">
-            <input name="oid" type="text" class="form-control" id="orderid" placeholder="Order ID" value="<?php if(isset($_GET['edit'])) echo $editrow['fld_order_id']; ?>" required>
-          </div>
+      </div>
+      <div class="form-group">
+        <label for="orderdate" class="col-sm-3 control-label">Order Date</label>
+        <div class="col-sm-9">
+          <input name="name" type="text" class="form-control" id="orderdate" placeholder="Order Date" value="<?php if(isset($_GET['edit'])) echo $editrow['fld_order_date']; ?>" required>
         </div>
-        <div class="form-group">
-          <label for="orderdate" class="col-sm-3 control-label">Order Date</label>
-          <div class="col-sm-9">
-            <input name="name" type="text" class="form-control" id="orderdate" placeholder="Order Date" value="<?php if(isset($_GET['edit'])) echo $editrow['fld_order_date']; ?>" required>
-          </div>
-        </div>
-        <div class="form-group">
-          <label for="orderstaff" class="col-sm-3 control-label">Staff</label>
-          <div class="col-sm-9">
-            <select name="sid" class="form-control" id="staffid" required>
+      </div>
+      <div class="form-group">
+        <label for="orderstaff" class="col-sm-3 control-label">Staff</label>
+        <div class="col-sm-9">
+          <select name="sid" class="form-control" id="staffid" required>
+            <?php
+            try {
+              $conn = new PDO("mysql:host=$servername;dbname=$dbname", $username, $password);
+              $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+              $stmt = $conn->prepare("SELECT * FROM tbl_staffs_a192212_pt2");
+              $stmt->execute();
+              $result = $stmt->fetchAll();
+            }
+            catch(PDOException $e){
+              echo "Error: " . $e->getMessage();
+            }
+            foreach($result as $staffrow) {
+              ?>
+              <?php if((isset($_GET['edit'])) && ($editrow['fld_staff_id']==$staffrow['fld_staff_id'])) { ?>
+                <option value="<?php echo $staffrow['fld_staff_id']; ?>" selected><?php echo $staffrow['fld_staff_name'];?></option>
+              <?php } else { ?>
+                <option value="<?php echo $staffrow['fld_staff_id']; ?>"><?php echo $staffrow['fld_staff_name'];?></option>
+              <?php } ?>
               <?php
-              try {
-                $conn = new PDO("mysql:host=$servername;dbname=$dbname", $username, $password);
-                $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
-                $stmt = $conn->prepare("SELECT * FROM tbl_staffs_a192212_pt2");
-                $stmt->execute();
-                $result = $stmt->fetchAll();
-              }
-              catch(PDOException $e){
-                echo "Error: " . $e->getMessage();
-              }
-              foreach($result as $staffrow) {
-                ?>
-                <?php if((isset($_GET['edit'])) && ($editrow['fld_staff_id']==$staffrow['fld_staff_id'])) { ?>
-                  <option value="<?php echo $staffrow['fld_staff_id']; ?>" selected><?php echo $staffrow['fld_staff_name'];?></option>
-                <?php } else { ?>
-                  <option value="<?php echo $staffrow['fld_staff_id']; ?>"><?php echo $staffrow['fld_staff_name'];?></option>
-                <?php } ?>
-                <?php
       } // while
       $conn = null;
       ?> 
@@ -128,8 +139,8 @@ include_once 'orders_crud.php';
       <tr>
         <th>Order ID</th>
         <th>Order Date</th>
-        <th>Staff ID</th>
-        <th>Customer ID</th>
+        <th>Staff</th>
+        <th>Customer</th>
         <th></th>
       </tr>
       <?php
@@ -146,7 +157,7 @@ include_once 'orders_crud.php';
         $sql = "SELECT * FROM tbl_orders_a192212_pt2, tbl_staffs_a192212_pt2, tbl_customers_a192212_pt2 WHERE ";
         $sql = $sql."tbl_orders_a192212_pt2.fld_staff_id = tbl_staffs_a192212_pt2.fld_staff_id and ";
         $sql = $sql."tbl_orders_a192212_pt2.fld_customer_id = tbl_customers_a192212_pt2.fld_customer_id";
-        $stmt = $conn->prepare("select * from tbl_orders_a192212_pt2 LIMIT $start_from, $per_page");
+        $stmt = $conn->prepare($sql . " LIMIT $start_from, $per_page");
         $stmt->execute();
         $result = $stmt->fetchAll();
       }
@@ -158,12 +169,14 @@ include_once 'orders_crud.php';
         <tr>
           <td><?php echo $orderrow['fld_order_id']; ?></td>
           <td><?php echo $orderrow['fld_order_date']; ?></td>
-          <td><?php echo $orderrow['fld_staff_id']; ?></td>
-          <td><?php echo $orderrow['fld_customer_id']; ?></td>
+          <td><?php echo $orderrow['fld_staff_name']; ?></td>
+          <td><?php echo $orderrow['fld_customer_name']; ?></td>
           <td>
             <a href="orders_details.php?oid=<?php echo $orderrow['fld_order_id']; ?>" class="btn btn-warning btn-xs" role="button">Details</a>
-            <a href="orders.php?edit=<?php echo $orderrow['fld_order_id']; ?>" class="btn btn-success btn-xs" role="button">Edit</a>
-            <a href="orders.php?delete=<?php echo $orderrow['fld_order_id']; ?>" onclick="return confirm('Are you sure to delete?');" class="btn btn-danger btn-xs" role="button">Delete</a>
+            <?php if ($role === 'ADMIN') {?>
+              <a href="orders.php?edit=<?php echo $orderrow['fld_order_id']; ?>" class="btn btn-success btn-xs" role="button">Edit</a>
+              <a href="orders.php?delete=<?php echo $orderrow['fld_order_id']; ?>" onclick="return confirm('Are you sure to delete?');" class="btn btn-danger btn-xs" role="button">Delete</a>
+            <?php }?>
           </td>
         </tr>
         <?php
